+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to this project will be documented in this file. The format 
 ### Added
 
 - `Spinner` component.
+- `languages` option to `generateRoutes` function, that allows to generate routes for multiple languages.
+- `InferRouteKey` type utility that infers the `RouteKey` type from the route object.
+- `useSwitchPathLang` hook, returned by `generatePath`. It allows to switch between languages keeping the language path synchronized.
+- Tests for routing.
+- `InferRouteKey` type utility that infers the `RouteKey` type from the route object.
 
 ### Changed
 
@@ -22,7 +27,7 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Fixed
 
--
+- Missing initial forward slash in `RouteKey`'s generated paths.
 
 ### Security
 
@@ -32,11 +37,16 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Added
 
-- `rightContent` prop to `Filters` component.
+- Added `CodeBlock` component.
+- Added `rightContent` prop to `Filters` component.
 
 ### Changed
 
 - Storybook bumped to v7
+
+### Fixed
+
+- Fixed router typing declarations.
 
 ## v0.1.0 - 27-03-2023